refactor(core): extract caret rect measurement into `caret-rect.ts`
